+++ test/pagination/keyset-paging.js
@@ -885,3 +885,47 @@ test('should handle an interface type', async t => {
   }
   t.deepEqual(expect, data.user.writtenMaterial)
 })
+
+test('should use default page limit if set', async t => {
+  const query = `{
+    users {
+      edges{
+        node{
+          id
+          comments {
+            edges{
+              node{
+                id
+              }
+            }
+          }
+        }
+      }
+    }
+  }`
+  const { data, errors } = await run(query)
+  errCheck(t, errors)
+  t.deepEqual(data.users.edges[0].node.comments.edges.length, 2)
+})
+
+test('should allow explicit page limit larger than default page size', async t => {
+  const query = `{
+    users {
+      edges{
+        node{
+          id
+          comments(first:3) {
+            edges{
+              node{
+                id
+              }
+            }
+          }
+        }
+      }
+    }
+  }`
+  const { data, errors } = await run(query)
+  errCheck(t, errors)
+  t.deepEqual(data.users.edges[0].node.comments.edges.length, 3)
+})